Fix BandMap scrollbar visibility and frequency formatting - v2.34.0
Fixed two issues:

1. BandMap scrollbars not appearing when window resized:
   - Added calculateColumnLayout() calls in addSpot(), removeSpot(), clearSpots()
   - Column count now updates when spots are added/removed
   - Horizontal scrollbar appears correctly when columns exceed viewport width

2. Debug message frequency formatting:
   - Changed from scientific notation (1.4034e+07) to integer format (14034000)
   - Cast frequency to qint64 for proper Hz display
   - Makes logs more readable: "DX Spot received: K3PR at 14034000 Hz"

Changes:
- BandMapWidget::addSpot() - call calculateColumnLayout() before updateScrollBars()
- BandMapWidget::removeSpot() - call calculateColumnLayout() before updateScrollBars()
- BandMapWidget::clearSpots() - call calculateColumnLayout() before updateScrollBars()
- MainWindow::onDXSpotReceived() - cast frequency to qint64 for qDebug()

+++ src/ui/MainWindow.cpp
@@ -1709,7 +1709,7 @@ void MainWindow::onDXSpotReceived(const QString& callsign,
                                    double frequency,
                                    const QString& spotter,
                                    const QString& comment) {
-    qDebug() << "DX Spot received:" << callsign << "at" << frequency << "Hz from" << spotter;
+    qDebug() << "DX Spot received:" << callsign << "at" << qint64(frequency) << "Hz from" << spotter;
 
     // If band map window exists, forward the spot to it
     if (m_bandMapWindow) {
